Fix circular references, auto-whitespaces, move helpers into dedicated folder and split few methods up
